Creating config.py and replace everything config data there
