+++ tools/bot.py
@@ -1,9 +1,3 @@
-import uvicorn
-
-from starlette.applications import Starlette
-from starlette.requests import Request
-from starlette.responses import PlainTextResponse, Response
-from starlette.routing import Route
 from telegram import Update
 from telegram.ext import Application, ApplicationBuilder, CommandHandler, CallbackContext
 
@@ -29,42 +23,16 @@ def add_handlers(app: Application) -> None:
             app.add_handler(handler(*params))
 
 
-async def init_webhook() -> None:
+async def init_webhook() -> Application:
+    """
+    Init webhook bot
+    """
     app = Application.builder().updater(None).token(config.TOKEN).build()
 
     add_handlers(app)
 
     await app.bot.set_webhook(url=f"{config.WEBHOOK_URL}/telegram")
-
-    async def telegram(request: Request) -> Response:
-        await app.update_queue.put(
-            Update.de_json(data=await request.json(), bot=app.bot)
-        )
-        return Response()
-
-    async def health(_: Request) -> PlainTextResponse:
-        return PlainTextResponse(content="The bot is still running fine :)")
-
-    starlette_app = Starlette(
-        routes=[
-            Route("/telegram", telegram, methods=["POST"]),
-            Route("/health", health, methods=["GET"]),
-        ]
-    )
-
-    webserver = uvicorn.Server(
-        config=uvicorn.Config(
-            app=starlette_app,
-            port=config.PORT,
-            use_colors=False,
-            host=config.HOST,
-        )
-    )
-
-    async with app:
-        await app.start()
-        await webserver.serve()
-        await app.stop()
+    return app
 
 
 def init_pooling():
